requirements.txt
pycocotools~=2.0.7
tqdm~=4.66.2
cython
scipy~=1.7.3
pandas~=1.3.5
seaborn~=0.12.2
matplotlib~=3.5.3
numpy~=1.21.6
MultiScaleDeformableAttention~=1.0
setuptools~=65.6.3
Pillow~=9.5.0
einops~=0.6.1
opencv-python~=4.1.2.30